requirements.txt
streamlit
pandas
PyMuPDF
docx2txt
python-docx
openpyxl